backend-services/airphynet-model/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pymongo==4.6.1
torch==2.1.2
numpy==1.26.3
python-dotenv==1.0.0
paho-mqtt==1.6.1
mlflow==2.10.0
dagshub==0.3.17
prometheus-fastapi-instrumentator==7.0.0